refactor multimodal analysis to own module
